Change create-repository folder structure and split create and list into 2 separate components
