+++ frontend/src/routes/logged-in-routes/workspaced/index.tsx
@@ -17,7 +17,8 @@ import RoleInfo from "~/pages/workspace/roles/role-info";
 import useFetchWorkspaces from "~/hooks/use-fetch/use-fetch-wokrspaces";
 import useFetchUserPermissions from "~/hooks/use-fetch/use-fetch-user-permissions";
 import CreateContainerRepository from "~/pages/container-repository/create";
-import ContainerInfo from "~/pages/container-repository/container-info";
+import ContainerInfo from "~/pages/container-repository/container";
+import ListContainerRepository from "~/pages/container-repository/list";
 
 const WorkspacedLayout = (props: ParentProps<{}>) => {
 	const [workspaces] = useFetchWorkspaces();
@@ -66,7 +67,8 @@ export default function WorkspacedRoutes() {
 				<Route path="/roles/new" component={CreateRoles} />
 			</Route>
 			<Route path="/container-repositories">
-				<Route path="/" component={CreateContainerRepository} />
+				<Route path="/" component={ListContainerRepository} />
+				<Route path="/new" component={CreateContainerRepository} />
 				<Route path="/:id" component={ContainerInfo} />
 			</Route>
 		</Route>
